databricks_notebook/buscador_medicamentos_app/lakehouse_app/requirements.txt
databricks-sdk
streamlit
pandas
geopy
speechrecognition
streamlit-geolocation
databricks-sql-connector
streamlit-audiorec
Unidecode
